requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic>=2
pydantic-settings
python-dotenv